Remove gosu/su-exec from entrypoint.sh
34ed7d5697edf8cfcd1af518c5764fceb3073bff removed the `FLUENT_UID`
environment variable to configure the fluent user, but did not
allow the container image to run as an arbitrary uid as the comment
linked from the commit message suggests. This leaves the
administrator with no way of configuring the identity of the user
the container runs as, which is necessary if mounting host volumes
into the container with restricted permissions.

Happily, `entrypoint.sh` no longer does anything that requires root
access, so by simply removing `gosu`/`su-exec`, using the `-u`
option with `docker run` just works. This also adds a `USER`
instruction so that the default configuration with no `-u` option
still works correctly.

Note that with this change, `entrypoint.sh.erb` no longer actually
contains any templating. This could be renamed to `entrypoint.sh`
and just symlinked to where it is used, but I'm not doing that here
to keep the diff straightforward.

Signed-off-by: Steven McDonald <steven.mcdonald@usabilla.com>